+++ client/views/Issue/IssuePageSingle.jsx
@@ -3,17 +3,21 @@ IssuePage = React.createClass({
     mixins: [ReactMeteorData],
 
     getMeteorData() {
+        var id = this.props.selected;
         Meteor.subscribe('contractors');
         Meteor.subscribe('services');
+        Meteor.subscribe('facilities');
         Meteor.subscribe('teamsAndFacilitiesForUser');
         Meteor.subscribe('users');
+        Meteor.subscribe('singleRequest',id);
         return {
-        	issue:Issues.findOne(this.props.selected)
+        	issue:Issues.findOne(id)
         }
     },
 
     render() {
-        if(!this.data.issue) {
+        var issue = this.data.issue;
+        if(!issue) {
             return <div/>
         }
     	return (
@@ -21,8 +25,11 @@ IssuePage = React.createClass({
                 <div className="issue-page wrapper wrapper-content animated fadeIn">
                     <div className="row">
                         <div className="col-xs-12">
+                            <h3>{issue.team.name}<br/>Work Request #{issue.code}</h3>
+                        </div>
+                        <div className="col-xs-12">
                             <div className="ibox">
-                            	<IssueDetail item={this.data.issue} />
+                            	<IssueDetail item={issue} />
                             </div>
                         </div>
                     </div>